add missing endpoints and refactor to look uniform
